torinometeo/requirements/common.txt
Django==1.8.5
django-getenv==1.3.1
Fabric==1.10.2
MySQL-python==1.2.5
Pillow==3.0.0
argparse==1.2.1
django-ckeditor==5.0.2
django-cleanup==0.4.0

django-disqus==0.5

django-pipeline==1.5.4
django-simple-captcha==0.4.6
django-taggit==0.18.0
ecdsa==0.13
funcsigs==0.4
futures==3.0.3
mock==1.3.0
paramiko==1.15.3
pbr==1.8.1
pycrypto==2.6.1
six==1.10.0
sorl-thumbnail==12.3
wsgiref==0.1.2

django-suit==0.2.15


Unidecode==0.04.18
django-filer==0.9.12
django-mptt==0.7.4
django-polymorphic==0.7.2
easy-thumbnails==2.2

Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
ansible==1.9.4
click==5.1
ordereddict==1.1
python-dotenv==0.1.5
